Refactor decay slightly
