+++ TFY4235_numfys/3/tests/qualitative/nuclear_decay.py
@@ -1,13 +1,14 @@
 """Run the nuclear decay simulation and plot it"""
 from nuclear import decay
+from nuclear.decay.utils import steps, half_life
 
 from matplotlib import pyplot as plt
 
 
 def plot_nuclear_decay(Ns0, lambdas, dt):
     Nsi = Ns0
-    n = decay.n(lambdas[0], 0.001, dt)
-    print(n)
+    n = steps(lambdas[0], 0.001, dt)
+
     t = [0]
     Ns = [[Nsi[0]], [Nsi[1]], [Nsi[2]]]
     for _ in range(n):
@@ -33,5 +34,5 @@ def main():
     lambdas = [10, 2]
     dt = 0.001
     for l in lambdas:
-        print(f"lambda: {l}, Half-life: {decay.half_life(l)}")
+        print(f"lambda: {l}, Half-life: {half_life(l)}")
     plot_nuclear_decay(Ns0, lambdas, dt)
